fix : back application, app sync
