+++ BE-WEB/discovery/src/main/resources/application-prd.yml
@@ -1,0 +1,8 @@
+eureka:
+  #  instance:
+  #    hostname: 3.35.136.193
+  client:
+    register-with-eureka: false
+    fetch-registry: false
+    service-url:
+      defaultZone: http://saveurlife.kr:8761/eureka
