+++ GoodNews/app/src/main/java/com/saveurlife/goodnews/sync/DataSync.kt
@@ -1,0 +1,4 @@
+package com.saveurlife.goodnews.sync
+
+class DataSync {
+}
